specialistAgent/requirements.txt
pydantic==2.10.6
flask==3.1.0
together==1.5.4
tiktoken==0.9.0
pinecone==6.0.2
python-dotenv==1.0.1